Rename `processId` to `processRef` to match the API
