+++ storage/DatabaseSchema.ts
@@ -1,5 +1,7 @@
 import { NamedSchema } from "remultiform/database/types";
 
+type ProcessRef = string;
+
 // We will replace this with a real name before release. It should be specific
 // to the environment it's running on to avoid clashes.
 export const databaseName = "mat-process-thc-local";
@@ -9,12 +11,12 @@ type DatabaseSchema = NamedSchema<
   1,
   {
     outsidePropertyImages: {
-      key: string;
+      key: ProcessRef;
       value: string[];
     };
 
     metalGateImages: {
-      key: string;
+      key: ProcessRef;
       value: string[];
     };
   }
